3-D rasteriser
Following the tutorial has accumulated in a serviceable 3D rasteriser, that can draw many triangles to the screen and can be viewed from all angles.
Triangles that are only partially on screen are clipped into smaller triangles, this can be seen by moving through the mountains.

The Utah teapot is used as a standard reference object, its sort of the hello world of 3d rendering.